extracted handlers and healthchecks into separate files because main is becoming tooooo big
